requirements.txt
cvzone==1.5.6
face_recognition==1.3.0
firebase_admin==6.1.0
Flask==2.2.3
matplotlib==3.5.2
numpy==1.24.2
opencv_python==4.7.0.72
Pillow==9.4.0
Werkzeug==2.2.3
